tdf#130857 qt weld: Add initial QtInstanceLabel implementation
Add a new QtInstanceLabel class as the native Qt
weld::Label implementation and implement the
straightforward and most important methods, and let
the less common ones trigger an assert for now.

Implement QtInstanceBuilder::weld_label to return
an instance of QtInstanceLabel.

This will e.g. be needed when adding support
for the "Tools" -> "Word Count" dialog in Writer
in the future, once more missing aspects have been
implemented.

Change-Id: I911bd68ad641b15883a0c6aa0707fd590fc05f96
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174366
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>