cool#9992 lok doc sign: avoid storing the sign cert in the model after sign
Open the signatures dialog in LOK mode, add a signature, close it, you
get a (non-async) popup asking if you want to automatically re-sign the
document on each save in the future.

This makes sense when only one user edits the document, but the LOK mode
stores the signing certificates in the view, so this would mean user A
can enable this mode, then user B could edit + save, which is not
wanted.

Fix the problem by just not offering this functionality in LOK mode,
where we assume that certificates are per-view and the save works on the
model.

Change-Id: I180fc2e6078623d44f8414f7891481218df7cebd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174369
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins