tdf#130857 qt weld: Move runAsync logic to QtInstanceDialog
Refactor the handling for running a QtInstanceMessageDialog
asynchronously, and reuse the logic for the QtInstanceDialog
base class:

Move implementations for both QtInstanceMessageDialog::runAsync
methods and the involved class members from QtInstanceMessageDialog
to QtInstanceDialog.

Split the previous logic from
QtInstanceMessageDialog::dialogFinished into two methods:

* move most of the logic as is to a new virtual slot
  in the base class, QtInstanceDialog::dialogFinished
* override the base class implementation to get the
  response code from the actually clicked button
  via the PROPERTY_VCL_RESPONSE_CODE property
  set on that button, and call the base class
  method with that one.

For QtInstanceDialog, there's no QDialog::clickedButton
method that could be used to retrieve the clicked button,
but it's also not needed there, because there a slot
connected to the QAbstractButton::clicked signal that triggers
closing the dialog, and passes the proper response code of the
corresponding button already, see

    commit 08b55df5c9e42c1ccb78a156261811875629342a
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Sep 28 00:32:15 2024 +0200

        tdf#130857 qt weld: Close dialog on button click

With this commit in place, QtInstanceDialog can now
also be run asynchronously, instead of the
QtInstanceDialog::runAsync methods
just returning true without doing anything.

This will be needed e.g. when adding support for the
"Tools" -> "Word Count" dialog in Writer.

Change-Id: I4edb9443cb11d1dc831a18f708cdbdd67c239aa6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174374
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins