slideshow: Fix "calculating" the hash for the object's paragraphs
Using the createEnumeration on the UNO object and finding the
UNO object that representing the paragraph, and using that pointer
as a unique identifier has doesn't work reliably. The reason for
this is probably that the createEnumeration call creates new UNO
objects that represent the paragraph, so the pointers change all
the time.

This changes the "calculation" of the paragraphs so that the hash
is actually combined using the hash of the object (which is always
of the XShape type and the hash is consistent with those) and the
paragraph index. This works well for the slideshow where we don't
edit the objects, but it can be a problem in the future if we will
allow editing.

Change-Id: I6dc9486a1e5d7c4ceb3d609613b60603886063c8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174381
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>