tdf#130857 qt weld: Add new layout to parent layout
Similar to the way it's already done for widgets
a few lines above, also add a newly created layout
to the parent layout (if the parent is a layout).

This makes the case where a QLayout is inside a
QLayout (e.g. a "GtkBox" inside a "GtkBox" in the
.ui file) work, where previously, the inner layout
wouldn't be added to the hierarchy, and thus
not be displayed in the dialog.

Quoting from the QLayout::addItem doc [1]:

> This function is not usually called in application code. To add a widget
> to a layout, use the addWidget() function; to add a child layout, use
> the addLayout() function provided by the relevant QLayout subclass.

Therefore, cast to QBoxLayout, the only QLayout
type used so far and call QBoxLayout::addLayout
to add the child layout.
(This might have to be extended when using e.g.
QGridLayout in the future as well.)

This will e.g. be needed for the "Tools" -> "Word count"
dialog in Writer or the "Insert" -> "Axes..." dialog
seen in Calc when a chart is selected -- once further
remaining aspects needed to be able to support those
have been implemented and they'll be enabled
in QtInstanceBuilder::IsUIFileSupported.

Change-Id: I518b9a33be1db419737a80aee44d93b2a7be1094
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174397
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins