tdf#163359: Fix duplication of TSA URLs on Signing Dialog
When using the Signing tab on the "Export as PDF" dialog, the TSA URLs will be
appended to the ComboBox each time a certificate is selected. Thus, if a user
selects a certificate and then another (or the same) is selected the ComboBox
will end up with duplicated values.
This patch moves the code to append the TSA URLs to the function that sets up
the signing page, fixing the problem.

Change-Id: I93d9cb61e10f18b0c33765d63a5ce47c92e588bf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174424
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit 12407f0668b042f5cbdffbddf65dc864c809d180)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174680
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>