tdf#163225 vcl: Erase listbox bg when redrawing on mouse move
Somehow, moving the listbox invalidation code from
ImplWin::PreNotify to ListBox::PreNotify in

    commit fd7cb42f7d17d03e4fac6d8c5f1d6c7c49a36fc6
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Feb 14 16:16:30 2023 +0100

        tdf#153520 vcl: Align listbox invalidation with mouseover check

triggered rendering artifacts for the tdf#163225 listbox example.
It seems that the previous rendering operation apparently seems a
pixel or so off from the current one and as the InvalidateFlags::NoErase
flag gets passed, what was drawn previously is not erased.

Don't pass that flag any more to ensure that the background is
properly erased, which fixes the rendering artifacts.

The flag's docmentation says:

    /** The invalidated area is painted with the background color/pattern. */
    NoErase              = 0x0004,

Dropping the flag doesn't negatively affect drawing a listbox
that has an explicit background color set in my tests, s. modified
sample document attachment 196871 in tdf#163225 .

Stop passing that flag in ListBox::StateChanged, too.

Change-Id: Ibcacbed01888449db062f84b4782feb9de69851c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174438
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>