tdf#130857 qt weld: Create QGroupBox for "GtkFrame"
Handle "GtkFrame" objects from .ui files and create
a QGroupBox [1] for them. While there's also a QFrame
class in Qt, QGroupBox seems more suitable, as it
allows setting a title, while QFrame doesn't.

Add special handling for the label child of the
"GtkFrame". Quoting from the GtkFrame doc [2]:

> The GtkFrame implementation of the GtkBuildable interface supports
> placing a child in the label position by specifying “label” as the
> “type” attribute of a <child> element. A normal content child can be
> specified without specifying a <child> type attribute.

Example for a frame label in chart2/uiconfig/ui/insertaxisdlg.ui:

        <child type="label">
          <object class="GtkLabel" id="label1">
            <property name="visible">True</property>
            <property name="can-focus">False</property>
            <property name="hexpand">True</property>
            <property name="label" translatable="yes" context="insertaxisdlg|label1">Axes</property>
            <attributes>
              <attribute name="weight" value="bold"/>
            </attributes>
          </object>
        </child>

In order to set the label's text as the text for the
QGroupBox, identify that label child in
QtBuilder::tweakInsertedChild, take over its
text as the QGroupBox title, and mark the label
for deletion, as it's no longer needed, since
the QGroupBox handles its title itself from
then on.

[1] https://doc.qt.io/qt-6/qgroupbox.html
[2] https://docs.gtk.org/gtk4/class.Frame.html

Change-Id: Ie917a501466bad8821fc3e7f5049db7c1a56995f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174465
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>