tdf#130857 qt weld: Declare support for chart "Insert Axes" dialog
Add the .ui file for the "Insert Axes" chart dialog
to the list of .ui files supported by QtInstanceBuilder.

Now that

    Change-Id: Ie917a501466bad8821fc3e7f5049db7c1a56995f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 4 10:38:32 2024 +0200

        tdf#130857 qt weld: Create QGroupBox for "GtkFrame"

added support for "GtkFrame", all widget types used
by that dialog are handled by QtBuilder.

This dialog can be seen as follows:

* start Calc
* "Insert" -> "Chart", confirm with "Finish" button
* "Insert" -> "Axes" (while the chart is still selected)

X and Y axes are shown/hidden as expected according
to whether the corresponding checkboxes are ticked
or unticked before closing the dialog using "OK".

Mnemonic underlines for the checkboxes are currently
not the way they're meant to be, but that's a different
issue that will be addressed in a separate commit.

Change-Id: I15901f83192ba33ddc5c5eb779be680f26cb6b55
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174466
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>