tdf#130857 qt weld: Handle help ID
Implement handling for the help ID by using a
new property PROPERTY_HELP_ID that is set
on the QtInstanceWidget's QWidget.

Implement QtInstanceWidget::set_help_id
and QtInstanceWidget::get_help_id accordingly.

For setting the ID, introduce and use a static
helper method QtInstanceWidget::setHelpId
and use that in QtBuilder::makeObject to
set the help ID based on help root and widget
ID, in line with what is done at the end of
VclBuilder::makeObject for the vcl::Window-based
builder.

Change-Id: I274886d8045b31ccbc92f586e2ead20ff7407d15
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174481
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins