tdf#130857 qt weld: Add basic handling for "GtkScrolledWindow"
Create a QScrollArea when encountering a
"GtkScrolledWindow" object in a .ui file.

This commit does not yet implement an
actual Qt-native weld::ScrolledWindow,
i.e. QtInstanceBuilder::weld_scrolled_window
currently still triggeres an assert.

This will help to support
cui/uiconfig/ui/objecttitledescdialog.ui,
which has a GtkScrolledWindow, but
SvxObjectTitleDescDialog doesn't explicitly
interact with it, so this basic implementation
is sufficient for a start.

Change-Id: I80d2565e57809be0cc3ae452359d0d1b827e46f6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174494
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins