tdf#119931: cui a11y: Specify mnenomic-widget in "Alt Text" dialog
Specify that the mnemonic widget for the "Alt Text"
label is the text view where the description can
be entered. This e.g. ensures that the a11y labelled-by
label-for relations get reported.

Drop the corresponding suppression for gla11y.

When using the Orca screen reader with the gtk3 VCL plugin,
Orca now announces "Alt Text" when the
text view gets focus, while that was not the case before.

Dialog can be triggered by e.g. selecting a QR code
in Writer, then opening the context menu and choosing
the "Alt Text" menu entry.

Change-Id: I5c8b06198a45cfbe5e1fc4a85ac568094b596819
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174506
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>