tdf#130857 VclBuilder: Move mnemonic-widget bookkeeping to base class
Move the `m_aMnemonicWidgetMaps` member and related struct
and typedef from VclBuilder's `m_pVclParserState` to the base
class BuilderBase's `m_pParserState` and also move the
VclBuilder::extractMnemonicWidget method to the base class,
so the logic can later be reused by QtBuilder.

Add a new getter BuilderBase::getMnemonicWidgetMaps to
get access and use that in VclBuilder instead.

At least for now, VclBuilder::extractMnemonicWidget
still explicitly needs to be called by VclBuilder
(and other subclasses) when encountering a "GtkLabel"
object for the mnemonic widget relationship to
be added to the map in the base class.

Change-Id: I878eec7be5e82fac3e1b944d7fed7bf6711744ce
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174508
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>