tdf#130857 qt a11y weld: Convert accelerator for labels
Convert the accelerator from GTK to Qt convention
for labels as well, to not display a literal "_",
but use the next letter as the accelerator.

For Qt, a literal "&" would still be displayed
if no buddy is actually set, i.e. there is no
related widget that can receive keyboard focus
when the accelerator key is pressed together with
Alt.

Setting buddies was implemented in previous commit

    Change-Id: Iba1e766c0f2c5162beb5b938c134afab0dac68b7
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 4 22:56:05 2024 +0200

        tdf#130857 qt a11y weld: Set "buddy" (mnemonic widget)

. Together with the above commit, this commit
makes the "Alt Text" label in the "Alt Text" dialog
(cui/ui/objecttitledescdialog.ui) show up as expected
with qt6 in a WIP branch that declares that .ui file
as supported in QtInstanceBuilder::IsUIFileSupported,
while it was previously shown as "_Alt Text".

Change-Id: Ief58dfc6864da9043766eec4eec36da44ae7017b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174511
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>