ScriptForge Fix tdf#163219 With blocks
Reference:
  https://bugs.documentfoundation.org/show_bug.cgi?id=163219#c7

Everywhere the With block variable is defined
or redefined inside the With block, the Basic
code has been reviewed.

Found in
  SFDatabases/SF_Database.xba/SetTransactionMode()
  SFDocuments/SF_Document.xba/Styles()
  ScriptForge/SF_Dictionary.xba/ImportFromPropertyValues()
  ScriptForge/SF_UI.xba/SetStatusBar()
  ScriptForge/SF_UI.xba/ShowProgressBar()

Most changes consist in isolating the With block
variable before the With block itself.

No functional change.

This solution is a workaround for the bug.
IT DOES NOT SOLVE THE ROOT CAUSE.

Change-Id: I48af29d3d9c8b1e36ef5a85c8cfe28f9639ae483
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174560
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins