a11y: Leave accessible description text view on simple tab key
Set the "accepts-tab" property to false for the GtkTextView
where an accessible description can be set, so pressing
the tab key moves focus to the next field in the dialog,
instead of inserting a literal tab character.

This simplifies using the dialog when using the keyboard only.
For the VCL and GTK implementations, pressing Ctrl+Tab
would previously be needed to move focus, which is less
intuitive.

Setting the property does not technically prevent inserting
tabs, they could still be inserted by copy-pasting
text containing tabs; just pressing the tab key while
the text view has focus won't insert them any more.
(I can't think of any reason to use actual tab characters
in the accessible description, however.)

Do this for the .ui files used for the following 2 scenarios:

* select QR code object in Writer, open context menu,
  select "Alt Text" context menu entry
* select an image in Writer, open context menu,
  select "Options" context menu entry, go to the
  "Options" tab page

Change-Id: I4bbecfc3f967b83768319e68ea3fb1f89b4c5014
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174569
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>