fix skia m117 sha value
so I messed up and uploaded skia m117 twice. But now there are two
copies of skia m117 in various CDN caches, with different sha256
output (no idea how that happened).

So upload a THIRD COPY with a new name, and then I can fix the
sha256 value in download.lst in a way that doesn't break the build for
someone.

Change-Id: I297ff4d9fa3a0edd902c85f2d674950945249a5e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174590
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>