avoid PDF export crash when ReadWithPossibleFixup falls back to Pdfium
Unfortunately, I failed to create a clean-room example,
even when I pasted some binary garbage at the end of the file.
Somehow I was getting a sanitized version from getAsStream.

In any case, this prevents crashing with the two examples I have,
although there are still significant problems somewhere
because the export just results in empty pages...

The problem is that reading again with the Pdfium-fixed version
gets mixed in with the original read results,
and things get pretty garbled.
Probably m_rElements and everything needs to be cleared
before tokenizing the fixup-version.

Change-Id: I01daefa23b8f92e2849d71bca269ba2b476199da
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174648
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>