PDFDocument::ReadWithPossibleFixup: clear before reading fixup
Ensure that a read of the pdfdocument
does not use info from failed previous reads.

The result in the problematic documents
was that GetCatalog failed because it was looking
for an m_aTrailerOffset[0] from the failed read,
and thus only empty pages were exported.

I failed to create a clean-room document,
and the single existing unit test that needs a fixup
has the trailer at the same GetPosition,
which replaced the m_aTrailerOffset
instead of creating a second entry.

Change-Id: I470b922323a2b2ebb6d4bbaa58c98f5cc492b750
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174652
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
(cherry picked from commit cb7a4d814644bf225127a93bbd6a114b6f19ac52)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174955
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>