tdf#134198 sw: avoid losing char pos of anchor point on content copy
Load the 5 pages bugdoc to Writer, notice the image on the last page,
select-all, copy, new document, paste the selection: no image on the
last page.

Given the new document may have different compat flags, this may not be
a bug, but bisect shows this was working before commit
591673d1637b25f83e01baccc321a83f10fc0ce7 (use more SwPosition::Assign,
2022-08-29). Also note that the image in question is anchored to-char
and it has a non-zero content (character) index, that's why it's on the
last page, and not on page 4. The old code kept the content index
unchanged while changing the node index, the new code avoids touching
the nNode member directly.

Fix the problem by still not touching the nNode member directly (to keep
the benefits of the above commit), but explicitly keep the content index
unchanged.

Changing the node index without changing the char index in general is
not safe (the new node may have less characters), but it does work in
this case, and this fixes the regression.

Change-Id: I2a90e4a88608d7435543880f46f09bea21eba48a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174655
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins