tdf#161479 make presentation work on fractionally scaled display
With fractional scaling on wayland the monitor sizes here are reported
effectively with the fractional scaling factor rounded up to the next
integer.

So:
1920 x 1080 at 125% scaling which appears like,
1536 x 864 is reported the same as 200% scaling, i.e.
960 x 540

So using these values as constraints for GDK_HINT_MAX_SIZE
with gtk_window_set_geometry_hints for m_bFullscreen are
too small to allow filling the screen.

Seeing as this was introduced as a workaround in the first
place, drop the workaround under wayland.

Change-Id: I7acfc929755728bbf77731d040f1f0f66cac836c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174761
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
(cherry picked from commit 4059c6351d9dc07ef2ab88149079cb2c59e22c32)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174783
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>