tdf#34804 sc: enable keyboard shortcut for CharColor and CharBackColor
A followup patch that adds keyboard shortcut support to Calc
for .uno:Color and .uno:CharBackColor to use recentColor.

Note that for Writer, it was .uno:FontColor (SID_ATTR_CHAR_COLOR2)
but the rest of the applications are .uno:Color (SID_ATTR_CHAR_COLOR)

I purposely did not add CharBackColor support to native Calc.
Notice that the "fill" icon on the toolbar is CELL-background, so the
color used would NOT NECESSARILY match what is seen in the toolbar,
and the sidebar also does not include a CharBackColor button.
[See patchset 1 for stuff needed for implementation.]

formatsh is for entire cells - .uno:Color opens dialog when no args: OK

editsc is for cursor-in-cell and for cursor-in-formula-bar.

drtxtobj is for comments, shapes, and textboxes (draw stuff).

Change-Id: I104b39eab1d475fc5c88837addf4bdbbd313d1be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174827
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins