a11y: Send value change event when checkbox value changes
VCLXAccessibleCheckBox implements XAccessibleValue. The value
returned by VCLXAccessibleCheckBox::getCurrentValue is based
on the current checkbox state and is 0 if the checkbox is
unchecked, 1 when it's checked, and 2 when it's indeterminate.

Therefore, also send a VALUE_CHANGED event when the value changes,
not just STATE_CHANGED events for the CHECKED and/or INDETERMINATE
states.

With this in place, changing the state of a checkbox (e.g.
in the "Tools" -> "Options" dialog) now causes the value displayed
in Accerciser's "Interface Viewer", section "Value" to be updated
right away when a checkbox is (un)checked when using the qt6 VCL
plugin, just as is the case for the "checked" state in the
"Accessible" section in Accerciser.
(Previously, the new value would only be shown after temporarily
selecting another object in the a11y tree, then selecting the
checkbox object again.)

Change-Id: I1a7da3b15a18beb12637a632c1779601ad1f198d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174829
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>