tdf#130857 VclBuilder: Move internal child visible logic to base
Move the special handling for internal children in .ui files
from VclBuilder to the base class and set the "visible" property
to true for them, instead of directly
calling vcl::Window::Show.

This allows to have this special logic that
internal children are visible by default, even
if the "visible" property is not explicitly set
in one place, and thus reuse it for QtBuilder as well.

Change-Id: Ic7a932556e02f47ee6007b5167f82bef152e1ef0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174833
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins