tdf#130857 qt weld: Set visibility based on "visible" prop
Use BuilderBase::extractVisible to assess and set the visibility
of a QWidget created from a .ui file.

As described in the GTK 3 documentation of the corresponding
GtkWidget property [1], the default value is false, i.e. if
an object is not explicitly set as visible, it should not be
shown.

This will be needed e.g. by the
svtools/uiconfig/ui/restartdialog.ui
dialog shown when toggling experimental mode in
"Tools" -> "Options" and closing the dialog with
the "OK" button. Without this preparatory commit,
using native Qt widgets would result in all the
labels (with different reasons for restarting)
being shown at the same time.

This depends on previous commit

    Change-Id: Ic7a932556e02f47ee6007b5167f82bef152e1ef0
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 11 23:01:44 2024 +0200

        tdf#130857 VclBuilder: Move internal child visible logic to base

to not cause buttons in dialogs to become invisible if
they're contained in the "internal" button box object
whose visibility is usually not explicitly set to True.

[1] https://docs.gtk.org/gtk3/property.Widget.visible.html

Change-Id: Iff254324a852a2fe73b3d9fcfe499bc8639380b6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174834
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins