tdf#121741 sd: not modified when leaving an empty placeholder
When you ENTER a placeholder, it removes the placeholder text.
In that case it doesn't treat this as a change
(because SdrObject::SetOutlinerParaObject is never called).

However, when you LEAVE the placeholder,
it puts the placeholder text back in.
This was being treated as a modification event (x2).

I tested what happens if you delete text
and then leave the placeholder.
The deleted text is already handled before we can leave the placeholder,
so that doesn't cause any concern.

TODO: an undo event is also created.

Change-Id: I366eb7a174284dc07415342258fcf65e50641be6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174860
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>