tdf#163364 package: ask to recover for this invalid ODF package
Bugdoc has a data descriptor on a folder entry, which is very odd and
entirely pointless.  Which is also the first entry, so it's an invalid
ODF package anyway.

ZipPackageFolder throws UnknownPropertyException for "WasEncrypted",
which results in General I/O error, but we want to ask the user if the
file should be opened in recovery mode.

(regression from commit 32cad89592ec04ab552399095c91dd76afb3002c)

Change-Id: Iafe610d507cf92d2fd2e9c3040592c3e638a30dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174889
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>