tdf#163397 svx: Use field color for WeldEditView bg
Use field color instead of the window color
for the WeldEditView background.

This is consistent with the color used for the other
text edit controls like the VCL Edit, see
Edit::ApplySettings.

While field and window color are the same color for e.g.
gtk3 or gen, they're different colors for macOS and
Qt-based VCL plugins, at least when using the Breeze
style that's default on KDE Plasma.

With this commit in place, the background color becomes
white for macOS and Linux with the qt5 VCL plugin with the
Breeze style.

For the latter, this had changed to light grey with

    commit 9dd26d7a2cc6b314dcc0b52000e0475bb4ced57d
    Date:   Thu Sep 5 01:37:44 2024 +0530

        Libreoffice Theme Part 0: Read correct colors from qt based VCL_PLUGINS

Change-Id: I42232fd07891046b18cc7f112118cac673e2090d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174890
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins