scriptforge.py Color args in exception.PythonShell()
The
   exception.PythonShell()
method opens the APSO console.

It receives 2 additional arguments
   background
   foreground
which specify the back- and foreground colors
of the window to open as a Python console.

Typical use of the method:
   exc.PythonShell({**globals(), **locals()}, \
       background = 0x0, foregound = 0xFFFFFF)
opens the shell with all existing variables
loaded and with white characters on a black
background.

The default values are those set in APSO:
   background = 0xFDF6E3
   foreground = 0x657B83
No impact on existing scripts.

The current change requires to be completed with
the addition of the 2 arguments in the
adequate help page.

Additonal minor changes
- typo in scriptforge.pyi
- 'outsideprocess' replaced by 'remoteprocess'

Change-Id: I9b1be123f049aa39bd0d814b6dd63fcc4f4ea310
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174897
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>