tdf#140061 CPPUNIT Test for fix swlo!sw::WriterMultiListener::StartListening
A CPPUNIT Test for the bug fix on swlo!sw::WriterMultiListener::StartListening
crash.
The sample file was added to the sw/qa/uiwriter/data.
Without this fix, writer crashes when the reproducing steps were followed.
The bug fix was reverted, when the test was ran, the test failed, i.e the
assert was not reached. The bug fix was reintroduced and the test
passed showing that the test catch the bug.

Change-Id: I688672d96a36d89c940bf0d401daaaeb732590a0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174908
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>