tdf#138711 sw textbox: capture fly when its draw object is captured
Load the bugdoc, notice that the only shape in the document has a
captured (inside the page frame) draw shape, but its inner fly frame is
not captured inside the page frame, so a small top margin is paired with
a larger right margin, which is unexpected.

This is nominally a regression from commit
9835a5823e0f559aabbc0e15ea126c82229c4bc7 (sw textboxes: reimplement ODF
import/export, 2014-10-04), because previously the ODF markup for Writer
TextBoxes were ignored, so we had correct position of the shape text at
the price of losing complex content.

Fix the problem similar to what commit
a0b6587c4acb1d74e1b00904147821640c98b323 (tdf#161199 sw
DoNotCaptureDrawObjsOnPage: capture wrap=none draw objects, 2024-06-13)
did: if we detect that the fly frame is part of a TextBox, then work
with the wrap type of the draw format, since these inner fly frames
always have their wrap type set to "through".

This fixes the unexpected position of the fly frame and keeps the import
of complex content working.

Change-Id: Ib0d861af0fac74d93a09178c10bc19a1b45e3ce7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/174987
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins